Refactor move commands into dedicated folders
