--- src/electionguard_cli/commands/__init__.py
@@ -1,11 +1,0 @@
-from electionguard_cli.commands import e2e_command
-from electionguard_cli.commands import import_ballots_command
-
-from electionguard_cli.commands.e2e_command import (
-    e2e,
-)
-from electionguard_cli.commands.import_ballots_command import (
-    import_ballots,
-)
-
-__all__ = ["e2e", "e2e_command", "import_ballots", "import_ballots_command"]
