+++ src/electionguard_cli/start.py
@@ -1,6 +1,6 @@
 import click
-from electionguard_cli.commands.e2e_command import e2e
-from electionguard_cli.commands.import_ballots_command import import_ballots
+from electionguard_cli.e2e.e2e_command import e2e
+from electionguard_cli.import_ballots.import_ballots_command import import_ballots
 
 
 @click.group()
